Develop to master

+++ files/default/solr-sync.sh
@@ -20,18 +20,24 @@ function set_dns_primary () {
 }
 
 function wait_for_replication_success () {
-  # wait up to 20 seconds for backup to complete, should only take a second or two
+  # Wait up to 30 seconds for backup to complete.
+  # Should only take a second or two for small indexes,
+  # but larger ones can be slow.
   BACKUP_STATUS=unknown
-  for i in {1..20}; do
+  for i in {1..30}; do
     if [ "$BACKUP_STATUS" = "unknown" ]; then
       DETAILS=$(curl "$HOST/$CORE_NAME/replication?command=details")
-      echo "$DETAILS" |grep 'status[^a-zA-Z]*success' && return 0
-      echo "$DETAILS" |grep 'exception.*snapshot' && return 1
+      echo "Backup status: $DETAILS"
+      if (echo "$DETAILS" |grep "$BACKUP_NAME"); then
+        echo "$DETAILS" |grep 'status[^a-zA-Z]*success' && return 0
+        echo "$DETAILS" |grep "exception.*$CORE_NAME" && return 1
+      fi
       sleep 1
     fi
   done
   if [ "$BACKUP_STATUS" = "unknown" ]; then
-    return 1
+    echo "Backup did not complete within 30 seconds"
+    return 2
   fi
 }
 
@@ -39,7 +45,10 @@ function export_snapshot () {
   # export a snapshot of the index and verify its integrity,
   # then copy to EFS so secondary servers can read it
   curl "$HOST/$CORE_NAME/replication?command=backup&location=$LOCAL_DIR&name=$BACKUP_NAME" | grep 'status[^a-zA-Z]*OK' || return 1
-  wait_for_replication_success || return 1
+  wait_for_replication_success; REPLICATION_STATUS=$?
+  if [ "REPLICATION_STATUS" != "0"]; then
+    return $REPLICATION_STATUS
+  fi
   sudo -u solr sh -c "$LUCENE_CHECK $LOCAL_SNAPSHOT && rsync -a --delete '$LOCAL_SNAPSHOT' '$SYNC_SNAPSHOT'" || return 1
 }
 
@@ -56,7 +65,14 @@ if (/usr/local/bin/pick-solr-master.sh); then
 
   # Export a snapshot of the index.
   # Drop this server from being master if it fails.
-  export_snapshot || (echo "" > $HEARTBEAT_FILE; exit 1) || exit 1
+  export_snapshot; EXPORT_STATUS=$?
+  if [ "$EXPORT_STATUS" != "0" ]; then
+    if [ "$EXPORT_STATUS" != "2" ]; then
+      echo "Export failed; assume server is unhealthy"
+      touch $HEARTBEAT_FILE.start
+    fi
+    exit 1
+  fi
 
   # clean up - remove old snapshots, hourly backup to S3
   for old_snapshot in $(ls -d $SYNC_DIR/snapshot.$CORE_NAME-* |grep -v "$SNAPSHOT_NAME"); do
